pyproject.toml
[tool.poetry]
name = "fastapi-media-stuff"
version = "0.1.0"
description = ""
authors = ["timelessnesses <mooping3roblox@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.79.1"
python-multipart = "^0.0.5"
pydub = "^0.25.1"
moviepy = "^1.0.3"
Pillow = "^9.2.0"
uvicorn = "^0.18.2"
sanitize-filename = "^1.2.0"
python-magic-bin = "^0.4.14"
python-magic = "^0.4.27"
fastapi-cache2 = "^0.1.9"
aioredis = "^2.0.1"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
autoflake = "^1.4"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
